refactor: modularize routing table logic into RoutingTableManager
